fix fm parser. directives: add `collection` and move them to `$lib/site`
The frontmatter parser was using a regex before, but this added complications when my `---(.*)---` regex had its wildcard match include actual markdown content up to a content string like `---`. Instead of chipping away at all the minutiae of excluding that sequence from the wildcard match, I refactored it into procedural code. It finds the first line of `---`, then finds the first line of `---` after that, then extracts the substring between the two. Come to think of it, I could accomplish the same with `regex.exec` and a regex of `^---$` where I run it twice and get the first two occurrences, but alas I've already implemented it this way.

The `collection` directive renders a list of links to all the pages in a collection.

I moved the directives to `$lib/site` to try and keep the `$lib/remark` as generic as possible.